Moving Form class to View\Components directory
