+++ config/components.php
@@ -2,7 +2,7 @@
 
 return [
 
-    \Styde\Form\Form::class => 'form',
+    \Styde\Form\View\Components\Form::class => 'form',
     \Styde\Form\View\Components\FormGroup::class => 'form-group',
     \Styde\Form\View\Components\Label::class => 'label',
     \Styde\Form\View\Components\Help::class => 'help',
